Aggiunto controllo su tutte le query

+++ 5_oral.php
@@ -12,21 +12,25 @@
     // Check if oral already exists
     $result_id_oral = $conn -> query("SELECT idoral FROM oral WHERE idstudent = $id_student AND idslot = $id_slot");
 
-    if ($result_id_oral -> num_rows > 0) {      // Oral remove
-        $row = $result_id_oral -> fetch_object();
-        $remove_oral = $conn -> query("DELETE FROM oral WHERE idoral = $row->idoral"); 
-        if ($remove_oral){
-            echo "Rimozione avvenuta con successo!";
-        } else {
-            die($conn->error); 
-        }    
-    } else {        // Oral insert
-        $insert_oral = $conn -> query("INSERT INTO oral VALUES ('', $id_student, $id_slot)");
-        if ($insert_oral){
-            echo "Inserimento avvenuto con successo!";
-        } else {
-            die($conn->error); 
+    if ($result_id_oral) {
+        if ($result_id_oral -> num_rows > 0) {      // Oral remove
+            $row = $result_id_oral -> fetch_object();
+            $remove_oral = $conn -> query("DELETE FROM oral WHERE idoral = $row->idoral"); 
+            if ($remove_oral){
+                echo "Rimozione avvenuta con successo!";
+            } else {
+                die($conn->error); 
+            }    
+        } else {        // Oral insert
+            $insert_oral = $conn -> query("INSERT INTO oral VALUES ('', $id_student, $id_slot)");
+            if ($insert_oral){
+                echo "Inserimento avvenuto con successo!";
+            } else {
+                die($conn->error); 
+            }
         }
+    } else {
+        die($conn->error);
     }
 
     // Close DB Connection
